+++ src/collective/dms/batchimport/utils.py
@@ -22,7 +22,15 @@ log = logging.getLogger("collective.dms.batchimport")
 
 
 def createDocument(
-    context, folder, portal_type, title, file_object, mainfile_type="dmsmainfile", owner=None, metadata=None
+    context,
+    folder,
+    portal_type,
+    title,
+    file_object,
+    mainfile_type="dmsmainfile",
+    owner=None,
+    metadata=None,
+    file_metadata=None,
 ):
     if owner is None:
         owner = api.user.get_current().id
@@ -55,10 +63,17 @@ def createDocument(
                 )
             )
 
-    file_title = _("Scanned Mail")
+    if not file_metadata:
+        file_metadata = {}
+    # title
     if "file_title" in metadata:
         file_title = metadata["file_title"]
         del metadata["file_title"]
+    if "title" not in file_metadata:
+        file_title = _("Scanned Mail")
+    file_metadata["title"] = file_title
+    # file
+    file_metadata["file"] = file_object
 
     with api.env.adopt_user(username=owner):
         document = createContentInContainer(folder, portal_type, **metadata)
@@ -67,6 +82,6 @@ def createDocument(
         if IDeadline and IDeadline.providedBy(document):
             document.deadline = deadlineDefaultValue(None)
 
-        version = createContentInContainer(document, mainfile_type, title=file_title, file=file_object)
+        version = createContentInContainer(document, mainfile_type, **file_metadata)
         log.info("file document has been created (id: %s)" % version.id)
         return (document, version)